Added multithreaded discovery functions
These gunctions are used to build a target list from IP ranges or Active
Directory lookups.  Multiple port scanning functions are available
including the one from PowerSploit.  I created by own
(Invoke-HuntPortScan) to optimize threading/performance for a small
number of ports/tests against a large number of hosts which is the
opposite requirement for vulnerability and service discovery (which
usually has a lot of tests for each host)